+++ src/radius/db/queries.py
@@ -31,5 +31,6 @@ QUERIES = {
     },
     "SELECT": {
         "AUTHOR": read("select_author"),
+        "TWEET": read("select_tweet"),
     },
 }