+++ src/radius/db/sql/select_tweet.sql
@@ -1,0 +1,9 @@
+SELECT
+    t.id,
+    a.display_name,
+    a.handle,
+    t.content,
+    t.created_at
+FROM tweets AS t
+JOIN authors AS a ON t.author_id = a.id
+WHERE t.id = ?;